Phase 4: decompose BlipLlmAssist into prompts, review logic, table, and hook
