+++ packages/client/src/components/radar/blipLlmReview.ts
@@ -1,0 +1,297 @@
+import type {
+  RadarBlip,
+  RadarQuadrant,
+  RadarRing,
+  TopicForTopicsPage,
+} from "@emstack/types";
+
+import { stripCodeFence } from "@/utils";
+
+export type Resolution
+  = | "create"
+    | "overwriteAll"
+    | "updateBlip"
+    | "removeBlip"
+    | "skip";
+
+export function isNoChangeSentinel(value: string): boolean {
+  const v = value.trim().toLowerCase();
+  return v === "no change" || v === "no-change" || v === "nochange";
+}
+
+export interface LlmEntry {
+  topic?: unknown;
+  quadrant?: unknown;
+  ring?: unknown;
+  description?: unknown;
+  radarNote?: unknown;
+  action?: unknown;
+}
+
+export interface EditDraft {
+  description: string;
+  radarNote: string;
+  quadrantId: string;
+  ringId: string;
+}
+
+export interface ResolvedLlmEntry {
+  topicName: string;
+  matchedTopicId: string | null;
+  willCreateTopic: boolean;
+
+  quadrantInput: string;
+  ringInput: string;
+
+  // Effective values (initially the LLM's, replaced by user edits)
+  description: string | null;
+  radarNote: string | null;
+  quadrantId: string | null;
+  ringId: string | null;
+
+  // Existing state at parse time
+  existingBlipId: string | null;
+  existingQuadrantId: string | null;
+  existingRingId: string | null;
+  existingRadarNote: string | null;
+  existingTopicDescription: string | null;
+
+  // Topic association counts (for delete-topic safety check)
+  topicCourseCount: number;
+  topicTaskCount: number;
+  topicDailyCount: number;
+
+  resolution: Resolution;
+  deleteTopicOnRemove: boolean;
+  editing: boolean;
+  editDraft: EditDraft | null;
+
+  selected: boolean;
+
+  problems: string[];
+}
+
+export function computeProblems(
+  entry: Pick<ResolvedLlmEntry,
+  "topicName" | "quadrantInput" | "quadrantId" | "ringInput" | "ringId"
+  | "resolution" | "existingBlipId">,
+  excludedNames?: Set<string>,
+): string[] {
+  const problems: string[] = [];
+  if (!entry.topicName) {
+    problems.push("missing topic");
+  }
+  else if (excludedNames?.has(entry.topicName.toLowerCase())) {
+    problems.push("topic is excluded from this radar");
+  }
+  // Skip and remove resolutions don't need quadrant/ring validation.
+  if (entry.resolution === "skip") {
+    return problems;
+  }
+  if (entry.resolution === "removeBlip") {
+    if (!entry.existingBlipId) {
+      problems.push("no existing blip to remove");
+    }
+    return problems;
+  }
+  if (entry.resolution === "updateBlip") {
+    if (!entry.existingBlipId) {
+      problems.push("no existing blip to update");
+    }
+    return problems;
+  }
+  if (!entry.quadrantId) {
+    problems.push(
+      entry.quadrantInput
+        ? `unknown slice "${entry.quadrantInput}"`
+        : "missing slice",
+    );
+  }
+  if (!entry.ringId) {
+    problems.push(
+      entry.ringInput ? `unknown ring "${entry.ringInput}"` : "missing ring",
+    );
+  }
+  return problems;
+}
+
+export function valuesEqual(a: string | null, b: string | null): boolean {
+  return (a ?? "") === (b ?? "");
+}
+
+export function descriptionChanged(r: ResolvedLlmEntry): boolean {
+  if (r.willCreateTopic) {
+    return false;
+  }
+  return !valuesEqual(r.description, r.existingTopicDescription);
+}
+
+export function radarNoteChanged(r: ResolvedLlmEntry): boolean {
+  if (!r.existingBlipId) {
+    return false;
+  }
+  return !valuesEqual(r.radarNote, r.existingRadarNote);
+}
+
+export function quadrantChanged(r: ResolvedLlmEntry): boolean {
+  if (!r.existingBlipId) {
+    return false;
+  }
+  return r.quadrantId !== r.existingQuadrantId;
+}
+
+export function ringChanged(r: ResolvedLlmEntry): boolean {
+  if (!r.existingBlipId) {
+    return false;
+  }
+  return r.ringId !== r.existingRingId;
+}
+
+function defaultResolution(args: {
+  isExcluded: boolean;
+  llmAction: string | null;
+  existingBlipId: string | null;
+}): Resolution {
+  if (args.isExcluded) {
+    return "skip";
+  }
+  if (args.llmAction === "remove" && args.existingBlipId) {
+    return "removeBlip";
+  }
+  if (args.llmAction === "update" && args.existingBlipId) {
+    return "overwriteAll";
+  }
+  if (args.llmAction === "add" && !args.existingBlipId) {
+    return "create";
+  }
+  return args.existingBlipId ? "overwriteAll" : "create";
+}
+
+export interface ParseLlmLookups {
+  topicByLowerName: Map<string, TopicForTopicsPage>;
+  quadrantByLowerName: Map<string, RadarQuadrant>;
+  ringByLowerName: Map<string, RadarRing>;
+  existingBlipByTopicId: Map<string, RadarBlip>;
+  excludedNamesLower: Set<string>;
+}
+
+export type ParseLlmEntriesResult
+  = | { entries: ResolvedLlmEntry[];
+    error: null; }
+    | { entries: null;
+      error: string; };
+
+export function parseLlmEntries(
+  jsonText: string,
+  lookups: ParseLlmLookups,
+): ParseLlmEntriesResult {
+  const {
+    topicByLowerName,
+    quadrantByLowerName,
+    ringByLowerName,
+    existingBlipByTopicId,
+    excludedNamesLower,
+  } = lookups;
+
+  let parsed: unknown;
+  try {
+    parsed = JSON.parse(stripCodeFence(jsonText));
+  }
+  catch (err) {
+    const message = err instanceof Error ? err.message : "Invalid JSON";
+    return {
+      entries: null,
+      error: message,
+    };
+  }
+  if (!Array.isArray(parsed)) {
+    return {
+      entries: null,
+      error: "Expected a JSON array.",
+    };
+  }
+
+  const entries = parsed as LlmEntry[];
+  const out: ResolvedLlmEntry[] = entries.map((entry) => {
+    const topicName = typeof entry.topic === "string" ? entry.topic.trim() : "";
+    const quadrantInput
+      = typeof entry.quadrant === "string" ? entry.quadrant.trim() : "";
+    const ringInput = typeof entry.ring === "string" ? entry.ring.trim() : "";
+    const llmAction = typeof entry.action === "string"
+      ? entry.action.trim().toLowerCase() || null
+      : null;
+
+    const quadrantMatch = quadrantInput
+      ? quadrantByLowerName.get(quadrantInput.toLowerCase())
+      : undefined;
+    const ringMatch = ringInput
+      ? ringByLowerName.get(ringInput.toLowerCase())
+      : undefined;
+    const topicMatch = topicName
+      ? topicByLowerName.get(topicName.toLowerCase())
+      : undefined;
+    const existingBlip = topicMatch
+      ? existingBlipByTopicId.get(topicMatch.id) ?? null
+      : null;
+
+    // "no change" sentinel preserves the existing value. null in JSON
+    // explicitly erases. Anything else is the new value.
+    const llmDescription = typeof entry.description === "string"
+      ? (isNoChangeSentinel(entry.description)
+        ? topicMatch?.description ?? null
+        : entry.description.trim() || null)
+      : null;
+    const llmRadarNote = typeof entry.radarNote === "string"
+      ? (isNoChangeSentinel(entry.radarNote)
+        ? existingBlip?.description ?? null
+        : entry.radarNote.trim() || null)
+      : null;
+
+    const isExcluded = excludedNamesLower.has(topicName.toLowerCase());
+    const resolution = defaultResolution({
+      isExcluded,
+      llmAction,
+      existingBlipId: existingBlip?.id ?? null,
+    });
+
+    const partial = {
+      topicName,
+      quadrantInput,
+      quadrantId: quadrantMatch?.id ?? null,
+      ringInput,
+      ringId: ringMatch?.id ?? null,
+      resolution,
+      existingBlipId: existingBlip?.id ?? null,
+    };
+
+    return {
+      topicName,
+      matchedTopicId: topicMatch?.id ?? null,
+      willCreateTopic: !!topicName && !topicMatch,
+      quadrantInput,
+      ringInput,
+      description: llmDescription,
+      radarNote: llmRadarNote,
+      quadrantId: quadrantMatch?.id ?? null,
+      ringId: ringMatch?.id ?? null,
+      existingBlipId: existingBlip?.id ?? null,
+      existingQuadrantId: existingBlip?.quadrantId ?? null,
+      existingRingId: existingBlip?.ringId ?? null,
+      existingRadarNote: existingBlip?.description ?? null,
+      existingTopicDescription: topicMatch?.description ?? null,
+      topicCourseCount: topicMatch?.resourceCount ?? 0,
+      topicTaskCount: topicMatch?.taskCount ?? 0,
+      topicDailyCount: topicMatch?.dailyCount ?? 0,
+      resolution,
+      deleteTopicOnRemove: false,
+      editing: false,
+      editDraft: null,
+      selected: false,
+      problems: computeProblems(partial, excludedNamesLower),
+    };
+  });
+  return {
+    entries: out,
+    error: null,
+  };
+}
